backend/requirements.txt
langchain
langchain-community
langchain-openai
langchain-text-splitters
faiss-cpu
pypdf
fastapi
uvicorn
python-dotenv
